Refactored code into seperate class files
