+++ src/ConnectFour.java
@@ -23,18 +23,18 @@ import java.awt.event.ActionListener;
 
 public class ConnectFour extends JPanel {
 
-    private enum GameStatus {RED_WON, YELLOW_WON, DRAW, RED_TURN, YELLOW_TURN};
+    public static enum GameStatus {RED_WON, YELLOW_WON, DRAW, RED_TURN, YELLOW_TURN};
 
-    private static final Font CONTROL_FONT = new Font("SansSerif", Font.BOLD, 16);
+    public static final Font CONTROL_FONT = new Font("SansSerif", Font.BOLD, 16);
 
-    private static final int ROWS = 6;
-    private static final int COLS = 7;
-    private static final int CONNECT = 4;
+    public static final int ROWS = 6;
+    public static final int COLS = 7;
+    public static final int CONNECT = 4;
 
     // Values to be used for button text within the board.
-    private static String RED = "R";
-    private static String YELLOW = "Y";
-    private static String EMPTY = " ";
+    public static String RED = "R";
+    public static String YELLOW = "Y";
+    public static String EMPTY = " ";
 
     private Icon redDiscIcon;
     private Icon yellowDiscIcon;
@@ -139,6 +139,14 @@ public class ConnectFour extends JPanel {
 
     public JLabel getStatusLabel() {
         return statusLabel;
+    }
+
+    public Icon getRedIcon() {
+        return redDiscIcon;
+    }
+
+    public Icon getYellowIcon() {
+        return yellowDiscIcon;
     }
     
     private void handleCellClick(int col) {
@@ -252,192 +260,5 @@ public class ConnectFour extends JPanel {
         }
 
         return GameStatus.DRAW;
-    }
-
-    /**
-     * COMPOSITE PATTERN: Individual cells and lines are treated uniformly by implementing GameComponent Interface
-     * checkWin() method returns false for leaves and for the composite checks if all children have the same value
-     * Cell value set to ClientProperty of corresponding JButton
-     */
-    interface GameComponent {
-        boolean checkWin();
-    }
-
-    class Cell implements GameComponent {
-        private Object value;
-
-        public Cell(Object value) {
-            this.value = value;
-        }
-
-        public Object getValue() {
-            return value;
-        }
-
-        public boolean checkWin() {
-            return false;
-        }
-    }
-
-    class LineComposite implements GameComponent {
-        private List<Cell> components;
-
-        public LineComposite(List<Cell> components) {
-            if (components.size() != CONNECT) { throw new IllegalArgumentException("Invalid Line Size"); }
-
-            this.components = components;
-        }
-
-        public boolean checkWin() {
-            Object first = components.get(0).getValue();
-            if (first.equals(EMPTY)) {
-                return false;
-            }
-
-            for (Cell c : components) {
-                if (!c.getValue().equals(first)) {
-                    return false;
-                }
-            }
-
-            return true;
-        }
-
-        public Object getFirstValue() {
-            return components.get(0).getValue();
-        }
-    }
-
-    /**
-     * TEMPLATE PATTERN: Abstract PlayerTurn provides template alogorithm for individual player turns. 
-     * RedPlayerTurn and YellowPlayer turn extend the base class with specific implementations
-     * Could be used to make a ComputerPlayer
-     */
-
-    abstract class PlayerTurn {
-        protected int selectedColumn;
-        protected ConnectFour game;
-
-        public PlayerTurn(ConnectFour game) {
-            this.game = game;
-        }
-
-        public final void takeTurn(int col) {
-            this.selectedColumn = col;
-
-            if (!dropDisc()) {
-                return;
-            }
-
-            switchTurns();
-            game.updateStatus();
-            
-        }
-
-        protected abstract boolean dropDisc();
-        protected abstract void switchTurns();
-    }
-
-    class RedPlayerTurn extends PlayerTurn {
-        RedPlayerTurn(ConnectFour game) {
-            super(game);
-        }
-
-        protected boolean dropDisc() {
-            for (int row = ConnectFour.ROWS - 1; row >= 0; row--) {
-                JButton cell = game.getCell(row, selectedColumn);
-                if (cell.getClientProperty("state").equals(EMPTY)) {
-                    cell.putClientProperty("state", RED);
-                    cell.setIcon(redDiscIcon);
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        protected void switchTurns() {
-            game.setRedTurn(false);
-        }
-    }
-
-    class YellowPlayerTurn extends PlayerTurn {
-        YellowPlayerTurn(ConnectFour game) {
-            super(game);
-        }
-
-        protected boolean dropDisc() {
-            for (int row = ConnectFour.ROWS - 1; row >= 0; row--) {
-                JButton cell = game.getCell(row, selectedColumn);
-                if (cell.getClientProperty("state").equals(EMPTY)) {
-                    cell.putClientProperty("state", YELLOW);
-                    cell.setIcon(yellowDiscIcon);
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        protected void switchTurns() {
-            game.setRedTurn(true);
-        }
-    }
-
-    /**
-     *  OBSERVER PATTERN: GameSubject manages a list of observers and notifies them when the game status changes
-     *  The GameObserver interface is implemented by different concrete observers
-     */
-    interface GameObserver {
-        void update(GameStatus status);
-    }
-
-    class GameSubject {
-        private List<GameObserver> observers = new ArrayList<GameObserver>();
-
-        public void addObserver(GameObserver observer) {
-            observers.add(observer);
-        }
-
-        public void notifyObservers(GameStatus status) {
-            for (GameObserver o : observers) {
-                o.update(status);
-            }
-        }
-    }
-
-    class StatusObserver implements GameObserver {
-        private JLabel statusLabel;
-
-        public StatusObserver(JLabel label) {
-            this.statusLabel = label;
-        }
-
-        public void update(GameStatus status) {
-            switch(status) {
-                case RED_WON:
-                    statusLabel.setText("Red Wins!");
-                    break;
-                case YELLOW_WON:
-                    statusLabel.setText("Yellow Wins!");
-                    break;
-                case DRAW:
-                    statusLabel.setText("Everyone is a winner!");
-                    break;
-                case RED_TURN:
-                    statusLabel.setText("Red\'s Turn.");
-                    break;
-                case YELLOW_TURN:
-                    statusLabel.setText("Yellow\'s Turn.");
-                    break;
-                default:
-                    statusLabel.setText("Invalid Status!");
-                    break;
-            }
-        }
-    }
-
-    class ConsoleObserver implements GameObserver {
-        public void update(GameStatus status) {
-            System.out.println("Game status updated to " + status);
-        }
     }
 }
