+++ src/ConsoleObserver.java
@@ -1,0 +1,9 @@
+package src;
+
+import src.ConnectFour.GameStatus;
+
+public class ConsoleObserver implements IGameObserver {
+    public void update(GameStatus status) {
+        System.out.println("Game status updated to " + status);
+    }
+}
